Tuan uwb1tag fix build

+++ UWB_2TAG_ANCHOR/Drivers/App/app_uart.h
@@ -14,7 +14,7 @@
 // TAG
 void app_uart_guide(UART_HandleTypeDef *huart, uint8_t *buffer);
 // Anchor
-void app_uart_dist(UART_HandleTypeDef *huart, uint8_t *buffer,uint8_t *tag_id);
+void app_uart_dist(UART_HandleTypeDef *huart, uint8_t *buffer,uint16_t *tag_id);
 void app_uart_item_orientation(UART_HandleTypeDef *huart, uint8_t *buffer);
 
 #endif /* APP_UART_APP_UART_H_ */
